pyproject.toml
[project]
name = "aliexpress-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "confluent-kafka>=2.12.2",
    "django>=6.0",
    "django-environ>=0.12.0",
    "django-prometheus>=2.4.0",
    "djangorestframework>=3.16.1",
    "elasticsearch>=9.2.0",
    "gunicorn>=23.0.0",
    "kafka-python>=2.3.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
